requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.104.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
ollama==0.4.7
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.11.2
pydantic-settings==2.0.3
pydantic_core==2.33.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.23.2
win32_setctime==1.2.0
